fix(headless): provide runtimeEnvironment for all Analytics Clients
https://coveord.atlassian.net/browse/KIT-1891
