+++ packages/headless/src/features/case-field/case-field-actions.ts
@@ -89,7 +89,7 @@ export const buildFetchClassificationRequest = async (
   url: state.configuration.platformUrl,
   caseAssistId: state.caseAssistConfiguration.caseAssistId,
   ...(state.configuration.analytics.enabled && {
-    visitorId: await getVisitorID(),
+    visitorId: await getVisitorID(state.configuration.analytics),
   }),
   fields: state.caseInput,
   context: state.caseField
